trainer-docker/trainer/requirements.txt
transformers==4.29.2
torch
numpy
pandas
seaborn
matplotlib
scikit-learn
google-cloud-storage